Added near-to-complete implementation for 1.17.1, added custom events to 1.8.9, added internal event managing system, added old utils back to 1.8.9, and started with universal Requisite menus.

+++ 1.8.9/src/main/java/xyz/matthewtgm/requisite/util/MouseHelper.java
@@ -24,11 +24,11 @@ import xyz.matthewtgm.requisite.core.util.IMouseHelper;
 
 public class MouseHelper implements IMouseHelper {
 
-    public int getMouseX() {
+    public double getMouseX() {
         return Mouse.getX() * UResolution.getScaledWidth() / UResolution.getWindowWidth();
     }
 
-    public int getMouseY() {
+    public double getMouseY() {
         return UResolution.getScaledHeight() - Mouse.getY() * UResolution.getScaledHeight() / UResolution.getWindowHeight() - 1;
     }
